+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `true`, `false`, and `NaN` are no longer parser constants [#292](https://github.com/jcubic/lips/issues/292)
 * get rid of node polyfills from output files
 * `eval` use `interaction-environment` stead of `current-environment` by default
+* remove support for string argument to `read` [#327](https://github.com/jcubic/lips/issues/327)
 ### Features
 * add `vector-for-each` and `vector-copy!` function from R7RS
 * add `string-for-each`, `string-downcase`, and `string-upcase` from R7RS